+++ src/__swaps__/screens/Swap/components/SearchInput.tsx
@@ -2,31 +2,26 @@ import { LIGHT_SEPARATOR_COLOR, SEPARATOR_COLOR, THICK_BORDER_WIDTH } from '@/__
 import { NavigationSteps, useSwapContext } from '@/__swaps__/screens/Swap/providers/swap-provider';
 import { opacity } from '@/__swaps__/utils/swaps';
 import { Input } from '@/components/inputs';
-import { AnimatedText, Bleed, Box, Column, Columns, Text, useColorMode, useForegroundColor } from '@/design-system';
+import { Bleed, Box, Column, Columns, Text, useColorMode, useForegroundColor } from '@/design-system';
 import * as i18n from '@/languages';
 import { userAssetsStore } from '@/state/assets/userAssets';
 import { useSwapsStore } from '@/state/swaps/swapsStore';
-import Clipboard from '@react-native-clipboard/clipboard';
 import React from 'react';
 import Animated, {
   runOnJS,
   runOnUI,
   useAnimatedProps,
   useAnimatedReaction,
-  useAnimatedStyle,
   useDerivedValue,
   useSharedValue,
 } from 'react-native-reanimated';
 import { useDebouncedCallback } from 'use-debounce';
-import { GestureHandlerV1Button } from './GestureHandlerV1Button';
+import { SearchInputButton } from './SearchInputButton';
 
 const AnimatedInput = Animated.createAnimatedComponent(Input);
 
 const FIND_A_TOKEN_TO_BUY_LABEL = i18n.t(i18n.l.swap.find_a_token_to_buy);
 const SEARCH_YOUR_TOKENS_LABEL = i18n.t(i18n.l.swap.search_your_tokens);
-const CANCEL_LABEL = i18n.t(i18n.l.button.cancel);
-const CLOSE_LABEL = i18n.t(i18n.l.button.close);
-const PASTE_LABEL = i18n.t(i18n.l.button.paste);
 
 export const SearchInput = ({
   handleExitSearchWorklet,
@@ -38,45 +33,11 @@ export const SearchInput = ({
   output: boolean;
 }) => {
   const { isDarkMode } = useColorMode();
-  const {
-    inputProgress,
-    inputSearchRef,
-    internalSelectedInputAsset,
-    internalSelectedOutputAsset,
-    outputProgress,
-    outputSearchRef,
-    AnimatedSwapStyles,
-  } = useSwapContext();
+  const { inputProgress, inputSearchRef, internalSelectedInputAsset, outputProgress, outputSearchRef } = useSwapContext();
 
   const fillTertiary = useForegroundColor('fillTertiary');
   const label = useForegroundColor('label');
   const labelQuaternary = useForegroundColor('labelQuaternary');
-
-  const btnText = useDerivedValue(() => {
-    if (
-      (inputProgress.value === NavigationSteps.SEARCH_FOCUSED && !output) ||
-      (outputProgress.value === NavigationSteps.SEARCH_FOCUSED && output)
-    ) {
-      return CANCEL_LABEL;
-    }
-
-    if ((output && internalSelectedOutputAsset.value) || !output) {
-      return CLOSE_LABEL;
-    }
-
-    return PASTE_LABEL;
-  });
-
-  const buttonVisibilityStyle = useAnimatedStyle(() => {
-    const isInputSearchFocused = inputProgress.value === NavigationSteps.SEARCH_FOCUSED;
-    const isVisible = isInputSearchFocused || output || (!output && internalSelectedInputAsset.value);
-
-    return {
-      display: isVisible ? 'flex' : 'none',
-      opacity: isVisible ? 1 : 0,
-      pointerEvents: isVisible ? 'auto' : 'none',
-    };
-  });
 
   const onInputSearchQueryChange = userAssetsStore(state => state.setSearchQuery);
 
@@ -111,17 +72,6 @@ export const SearchInput = ({
       }
     }
   );
-
-  const onPaste = () => {
-    Clipboard.getString().then(text => {
-      // to prevent users from mistakingly pasting long ass texts when copying the wrong thing
-      // we slice the string to 42 which is the size of a eth address,
-      // no token name query search should be that big anyway
-      const v = text.trim().slice(0, 42);
-      pastedSearchInputValue.value = v;
-      useSwapsStore.setState({ outputSearchQuery: v });
-    });
-  };
 
   return (
     <Box paddingHorizontal="20px" width="full">
@@ -189,54 +139,12 @@ export const SearchInput = ({
           </Bleed>
         </Box>
         <Column width="content">
-          <Animated.View style={buttonVisibilityStyle}>
-            <GestureHandlerV1Button
-              onPressJS={() => {
-                (output ? outputSearchRef : inputSearchRef).current?.blur();
-              }}
-              onPressWorklet={() => {
-                'worklet';
-                if (output && outputProgress.value === NavigationSteps.TOKEN_LIST_FOCUSED && !internalSelectedOutputAsset.value) {
-                  runOnJS(onPaste)();
-                }
-
-                if (
-                  isSearchFocused.value ||
-                  (output && internalSelectedOutputAsset.value) ||
-                  (!output && internalSelectedInputAsset.value)
-                ) {
-                  handleExitSearchWorklet();
-                }
-              }}
-              scaleTo={0.8}
-            >
-              <Box
-                as={Animated.View}
-                alignItems="center"
-                borderRadius={18}
-                height={{ custom: 36 }}
-                justifyContent="center"
-                paddingHorizontal={{ custom: 12 - THICK_BORDER_WIDTH }}
-                style={[
-                  output ? AnimatedSwapStyles.searchOutputAssetButtonWrapperStyle : AnimatedSwapStyles.searchInputAssetButtonWrapperStyle,
-                  {
-                    borderCurve: 'continuous',
-                    borderWidth: THICK_BORDER_WIDTH,
-                    overflow: 'hidden',
-                  },
-                ]}
-              >
-                <AnimatedText
-                  align="center"
-                  style={output ? AnimatedSwapStyles.searchOutputAssetButtonStyle : AnimatedSwapStyles.searchInputAssetButtonStyle}
-                  size="17pt"
-                  weight="heavy"
-                >
-                  {btnText}
-                </AnimatedText>
-              </Box>
-            </GestureHandlerV1Button>
-          </Animated.View>
+          <SearchInputButton
+            output={output}
+            pastedSearchInputValue={pastedSearchInputValue}
+            isSearchFocused={isSearchFocused}
+            handleExitSearchWorklet={handleExitSearchWorklet}
+          />
         </Column>
       </Columns>
     </Box>
